Move BlockNumberAndHash to ckb-types
