+++ sync/src/synchronizer/block_fetcher.rs
@@ -1,5 +1,5 @@
 use crate::block_status::BlockStatus;
-use crate::types::{ActiveChain, BlockNumberAndHash, HeaderIndex, HeaderIndexView, IBDState};
+use crate::types::{ActiveChain, HeaderIndex, HeaderIndexView, IBDState};
 use crate::SyncShared;
 use ckb_constant::sync::{
     BLOCK_DOWNLOAD_WINDOW, CHECK_POINT_WINDOW, INIT_BLOCKS_IN_TRANSIT_PER_PEER,
@@ -9,6 +9,7 @@ use ckb_logger::{debug, trace};
 use ckb_network::PeerIndex;
 use ckb_systemtime::unix_time_as_millis;
 use ckb_types::packed;
+use ckb_types::BlockNumberAndHash;
 use std::cmp::min;
 use std::sync::Arc;
