+++ sync/src/tests/inflight_blocks.rs
@@ -1,7 +1,8 @@
-use crate::types::{BlockNumberAndHash, InflightBlocks};
+use crate::types::InflightBlocks;
 use ckb_constant::sync::BLOCK_DOWNLOAD_TIMEOUT;
 use ckb_types::h256;
 use ckb_types::prelude::*;
+use ckb_types::BlockNumberAndHash;
 use std::collections::HashSet;
 
 #[test]
